+++ src/main/java/com/basketbandit/core/modules/game/commands/CommandWorldOfWarcraftCharacter.java
@@ -154,8 +154,7 @@ public class CommandWorldOfWarcraftCharacter extends Command {
             Utils.sendMessage(e, embed.build());
 
         } catch(Exception ex) {
-            Utils.sendMessage(e, "There was an issue processing your request.");
-            ex.printStackTrace();
+            Utils.sendMessage(e, "There was an issue processing the request for command: " + e.getMessage().getContentDisplay());
         }
 
     }