+++ src/main/java/com/basketbandit/core/Configuration.java
@@ -34,8 +34,11 @@ public class Configuration {
     // Bot's global command prefix.
     public static String GLOBAL_PREFIX;
 
-    // WOW API key.
+    // WoW API key.
     public static String WOW_API;
+
+    // Osu API Key
+    public static String OSU_API;
 
     // Bot's status, e.g Playing with lottie's tits.
     static String STATUS = "@BasketBandit help";